whole whack of changes, renames, removals, refactors, etc.
