+++ Library/sources/gumroad-products.js
@@ -1,5 +1,7 @@
 import 'dotenv/config'
 
+// import simple click though shop, via gumroad.
+// get all products
 export default async function GumRoadProducts() {
   try {
     const url = `https://api.gumroad.com/v2/products?access_token=${process.env.GUMROAD}`
