--- app/markdown/about.md
@@ -1,4 +1,0 @@
-[//]: # ({"title": "about", "creation": "01/01/1970", "summary": "about the creator of this webspace", "thumbnail": "/images/image.webp"})
-
-# about/valaxin
-
