--- app/markdown/example-article copy.md
@@ -1,3 +1,0 @@
-[//]: # ({"title": "Nothing Post Title", "creation": "01/03/2024", "summary": "nothing post for testing also "})
-
-# Nothing Post Title sadasdas
