+++ app/modules/class-injections.js
@@ -1,13 +1,12 @@
 'use strict'
 
-// adds 'bulma' classes to custom containers within a given blog post
-function articleContainers(articleSelector) {
-
+// adds bulma style classes to custom containers within a given blog post
+const customArticleContainers = (articleSelector) => {
   try {
     let elements = ''
     const elementTypes = ['spoilers', 'warning', 'information']
     const articleBody = document.body.querySelectorAll(articleSelector)[0]
-    if (!articleBody) { 
+    if (!articleBody) {
       console.log('no article')
       return false
     }
@@ -26,31 +25,27 @@ function articleContainers(articleSelector) {
   }
 }
 
-// add highlighting classes to current page within navigation elements
-function navHighlighting(navigationSelector) {
+// add highlighting classes to current page anchor within <nav> element
+const navHighlighting = (navigationSelector) => {
   let navigationLinks = document.body.querySelectorAll(navigationSelector)
-  console.log(window.location.pathname, navigationLinks)
-
+  // console.log(window.location.pathname, navigationLinks)
   for (link of navigationsLinks) {
-    
+    // console.log(link)
   }
 }
 
+// self calling anonymous call
 // add bulma classes when viewport reaches a defined width size
-
 export default (async function () {
   let pathSlug = 'blog'
   if (window.location.pathname.includes(`/${pathSlug}`)) {
-    
     const containers = articleContainers('section.content > article.content')
     if (containers) {
       console.log('custom-containers:', containers)
     }
   }
-
   const links = navHighlighting('nav.navbar > div a')
   if (links) {
     console.log('navigation-links', links)
   }
-
 })()
