--- app/templates/blog.pug
@@ -1,5 +1,0 @@
-extends partials/layout
-
-block content
-  include partials/directory
-  include partials/footer
