+++ app/templates/partials/codepen-projects.pug
@@ -1,0 +1,32 @@
+// h2.showcased-section-title ✨ 
+.codepen-projects
+  .grid.is-col-min-12
+    // grid of three items...
+    a(href="https://codepen.io/valaxin/details/AXMbyX" target="_blank")
+      .cell
+        h3.title Basic CSS3 Slider
+        p Just a basic slider with button controls on the bottom
+        .tags
+          .tag May 27, 2013
+          .tag 413 Loves
+          .tag 210930 Views
+
+    a(href="https://codepen.io/valaxin/details/BjepPG" target="_blank")
+      .cell
+        h3.title Another Toggle
+        p simple bool toggle
+        .tags
+          .tag February 18, 2016
+          .tag 147 Loves
+          .tag 25894 Views
+        p.is-small view on codepen.io
+          i.fa-fas-save
+
+    a(href="https://codepen.io/valaxin/details/XdBKBe" target="_blank")
+      .cell
+        h3.title Simple Check
+        p small checkmark styled to look pleasing and provide clear feedback
+        .tags
+          .tag April 21, 2016
+          .tag 42 Loves
+          .tag 2774 View
