--- app/templates/shop.pug
@@ -1,8 +1,0 @@
-extends partials/layout
-
-block content
-  include partials/products
-
-  section.products
-    h1.title Affliate Links Are Below Here [todo!]
-  include partials/footer
