--- library/data/sitemap.js
@@ -1,6 +1,0 @@
-export default function (indexed) {
-  console.log(indexed)
-  return {
-    "nothing": "here"
-  }
-}
